feat: base script tag to handle generation of queries considering main text types
It is possible to inform either CHART, TEXT, VARCHAR or ENUM and get insert rows of strings with random characters based on some of their limits (some because not everything about them has been enforced, such as the fact that CHAR can at max have 255 characters) determined by the user or pre-determined (ex: TEXT always has the same limit: 65535 characters). For now the insert is only being output in the console.